species_lists: move 3 forms to Views/
Per #4331:
- Components::SpeciesListForm
    -> Views::Controllers::SpeciesLists::Form
- Components::FormListFeedback
    -> Views::Controllers::SpeciesLists::WriteIn::ListFeedback
    (nested under write_in subnamespace — only caller is
    species_lists/write_in/_form.html.erb)
- Components::PrintLabelsForm
    -> Views::Controllers::SpeciesLists::Downloads::Form
    (nested under downloads subnamespace — only caller is
    species_lists/downloads/new.html.erb)

Components::FormLocationFeedback **stays in app/components/** —
genuinely reusable across species_lists, locations, AND
observations controllers (the audit had it as single-use to
species_lists but it's actually rendered from LocationForm,
ObservationFormDetails, SpeciesListForm, and the write_in partial).

4 ERB callers updated. 66 view + controller tests + 3 integration
tests pass. Lint clean.

Also tightened some forbidden `text:` regex assertions in the
moved list_feedback test (per .claude/rules/testing.md — every PR
that touches a test file is the right place to clean up forbidden
patterns).
