+++ app/views/controllers/species_lists/new.html.erb
@@ -4,7 +4,7 @@ add_context_nav(species_list_form_new_tabs)
 container_class(:text)
 %>
 
-<%= render(Components::SpeciesListForm.new(
+<%= render(Views::Controllers::SpeciesLists::Form.new(
              @species_list,
              projects: @projects,
              dubious_where_reasons: @dubious_where_reasons,
